Separate sanction check config DTO into discrete file.
