+++ dto/scenario_iterations.go
@@ -7,7 +7,6 @@ import (
 	"github.com/cockroachdb/errors"
 
 	"github.com/checkmarble/marble-backend/models"
-	"github.com/checkmarble/marble-backend/utils"
 )
 
 // Read DTO
@@ -33,103 +32,6 @@ type ScenarioIterationBodyDto struct {
 	ScoreRejectThreshold_deprec   *int                 `json:"score_reject_threshold"` //nolint:tagliatelle
 	ScoreDeclineThreshold         *int                 `json:"score_decline_threshold"`
 	Schedule                      string               `json:"schedule"`
-}
-
-type SanctionCheckConfig struct {
-	Datasets      []string                  `json:"datasets,omitempty"`
-	ForceOutcome  *string                   `json:"force_outcome,omitempty"`
-	ScoreModifier *int                      `json:"score_modifier,omitempty"`
-	TriggerRule   *NodeDto                  `json:"trigger_rule"`
-	Query         *SanctionCheckConfigQuery `json:"query"`
-}
-
-func AdaptSanctionCheckConfig(model models.SanctionCheckConfig) (SanctionCheckConfig, error) {
-	nodeDto, err := AdaptNodeDto(model.TriggerRule)
-	if err != nil {
-		return SanctionCheckConfig{}, nil
-	}
-
-	query, err := AdaptSanctionCheckConfigQuery(model.Query)
-	if err != nil {
-		return SanctionCheckConfig{}, err
-	}
-
-	config := SanctionCheckConfig{
-		Datasets:      model.Datasets,
-		ForceOutcome:  model.Outcome.ForceOutcome.MaybeString(),
-		ScoreModifier: &model.Outcome.ScoreModifier,
-		TriggerRule:   &nodeDto,
-		Query:         &query,
-	}
-
-	return config, nil
-}
-
-func AdaptSanctionCheckConfigInputDto(dto SanctionCheckConfig) (models.UpdateSanctionCheckConfigInput, error) {
-	config := models.UpdateSanctionCheckConfigInput{
-		Datasets: dto.Datasets,
-		Outcome: models.UpdateSanctionCheckOutcomeInput{
-			ScoreModifier: dto.ScoreModifier,
-		},
-	}
-
-	if dto.TriggerRule != nil {
-		astRule, err := AdaptASTNode(*dto.TriggerRule)
-		if err != nil {
-			return models.UpdateSanctionCheckConfigInput{}, errors.Wrap(
-				models.BadParameterError,
-				"invalid trigger",
-			)
-		}
-		config.TriggerRule = &astRule
-	}
-
-	if dto.Query != nil {
-		query, err := AdaptSanctionCheckConfigQueryDto(*dto.Query)
-		if err != nil {
-			return models.UpdateSanctionCheckConfigInput{}, errors.Wrap(
-				models.BadParameterError,
-				"invalid query",
-			)
-		}
-
-		config.Query = &query
-	}
-	if dto.ForceOutcome != nil {
-		config.Outcome.ForceOutcome = utils.Ptr(models.ForcedOutcomeFrom(*dto.ForceOutcome))
-	}
-
-	return config, nil
-}
-
-type SanctionCheckConfigQuery struct {
-	Name NodeDto `json:"name"`
-}
-
-func AdaptSanctionCheckConfigQuery(model models.SanctionCheckConfigQuery) (SanctionCheckConfigQuery, error) {
-	nameAst, err := AdaptNodeDto(model.Name)
-	if err != nil {
-		return SanctionCheckConfigQuery{}, err
-	}
-
-	dto := SanctionCheckConfigQuery{
-		Name: nameAst,
-	}
-
-	return dto, nil
-}
-
-func AdaptSanctionCheckConfigQueryDto(dto SanctionCheckConfigQuery) (models.SanctionCheckConfigQuery, error) {
-	nameAst, err := AdaptASTNode(dto.Name)
-	if err != nil {
-		return models.SanctionCheckConfigQuery{}, err
-	}
-
-	model := models.SanctionCheckConfigQuery{
-		Name: nameAst,
-	}
-
-	return model, nil
 }
 
 func AdaptScenarioIterationWithBodyDto(si models.ScenarioIteration) (ScenarioIterationWithBodyDto, error) {
@@ -170,7 +72,7 @@ func AdaptScenarioIterationWithBodyDto(si models.ScenarioIteration) (ScenarioIte
 			Query:         &queryDto,
 		}
 
-		if si.SanctionCheckConfig.Outcome.ForceOutcome != models.Approve {
+		if si.SanctionCheckConfig.Outcome.ForceOutcome != models.UnsetForcedOutcome {
 			body.SanctionCheckConfig.ForceOutcome =
 				si.SanctionCheckConfig.Outcome.ForceOutcome.MaybeString()
 		}
